+++ src/Widgets/Table.php
@@ -20,6 +20,7 @@ use Konekt\AppShell\Contracts\Theme;
 use Konekt\AppShell\Contracts\Widget;
 use Konekt\AppShell\Traits\RendersThemedWidget;
 use Konekt\AppShell\Widgets\Table\Columns;
+use Konekt\AppShell\Widgets\Table\Footer;
 
 class Table implements Widget
 {
@@ -31,12 +32,15 @@ class Table implements Widget
 
     public array $options = [];
 
+    public ?Footer $footer;
+
     public function __construct(Theme $theme, array $columns = [], array $options = [])
     {
         $this->theme = $theme;
         $this->columns = new Columns($columns);
         $this->options = $options;
         $this->data = collect([]);
+        $this->footer = isset($options['footer']) ? new Footer($options['footer']) : null;
     }
 
     public static function create(Theme $theme, array $options = []): Table
@@ -47,6 +51,11 @@ class Table implements Widget
     public function rendersAlternativeForEmptyDataset(): bool
     {
         return !empty($this->options['empty']['text'] ?? '');
+    }
+
+    public function hasFooter(): bool
+    {
+        return null !== $this->footer;
     }
 
     public function render($data = null): string